% Generated by roxygen2 (4.0.2): do not edit by hand
\name{narrow_until_normal}
\alias{narrow_until_normal}
\title{narrow_until_normal}
\usage{
narrow_until_normal(x, w = NA, min.pval = 0.05, min.left = 0,
  low.noise = 1/3, test = "shapiro")
}
\arguments{
\item{x}{array of input values or measurements}

\item{w}{array of weights}

\item{min.pval}{threshold for test's p-value}

\item{min.left}{fraction of x to keep (between 0 and 1; defaults to 0)}

\item{low.noise}{stop flagging outliers if sd falls below this}

\item{test}{name of the test to perform ('shapiro' or 'ks')}
}
\description{
This function uses a test for standard-distribution (defaults to shapiro)
and 'masks' outliers (largest deviation from weighted mean) until the
remaining data passes the specified test (follows normal distribution).
An array of 0 (usable) and 1 (flagged) will be returned.
}

