% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quandico}
\alias{quandico}
\title{quandico}
\usage{
quandico(sample = NA, sample.x = 2, sample.y = 0, reference = NA,
  reference.x = 2, reference.y = 0, sample.columns = c("chr_full",
  "position", "direction", "base", "gene", "sample", "orgname"),
  reference.columns = c("chr_full", "position", "direction", "base", "gene",
  "reference", "orgname"), output.dir = "..",
  output.basename = "quandico_run",
  assembly.filename = config.assembly.filename,
  assembly.version = config.assembly.version)
}
\arguments{
\item{sample}{the full path to the count file containing sample data (NA)}

\item{sample.x}{the number of X chromosomes in the sample (2)}

\item{sample.y}{the number of Y chromosomes in the sample (0)}

\item{reference}{the full path to the count file containing reference data}

\item{reference.x}{the number of X chromosomes in the reference (2)}

\item{reference.y}{the number of Y chromosomes in the reference (0)}

\item{sample.columns}{column definition for the sample input file (NA)
("chr_full", "position", "direction", "base", "gene", "sample", "orgname")}

\item{reference.columns}{column definition for the reference input file
("chr_full", "position", "direction", "base", "gene", "reference", "orgname")}

\item{output.dir}{the full path where to write results to}

\item{output.basename}{prefix for output files}

\item{assembly.filename}{file containing chromosome details (NA = hg19.fa.fai)}

\item{assembly.version}{name of the underlying genome build (NA = hg19)}
}
\description{
This is the main function of the package which will process read-count data
and perform copy-number estimation. Results will be written to four output
files:
\itemize{
\item{1. <basename>.csv}{comma separated values (CSV) and header}
\item{2. <basename>.vcf}{VCF file with header}
\item{3. <basename>.pdf}{PDF file with a graphical overview for every cluster}
\item{4. <basename>_overview.pdf}{summary plot with one point-range per cluster}
}
}
\examples{
\dontrun{
quandico(
	sample            = "/path/to/sample_counts_file.tab",
	reference         = "/path/to/reference_counts_file.tab",
	assembly.filename = "/path/to/fasta_index/hg19.fa.fai",
	assembly.version  = "hg19",
	output.dir        = "..",
	output.basename   = "sample_vs_ref"
)
}
}

