% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quandico.R
\name{var_wt}
\alias{var_wt}
\title{var_wt}
\usage{
var_wt(x, w, mu = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{array of input values or measurements}

\item{w}{array of weights}

\item{mu}{optional: weighted mean of values (x) with weights (w)}

\item{na.rm}{remove undefined values and weights (TRUE)}
}
\description{
This function will return the weighted variance of an array of values (x)
with associated weights (w). If the weighted mean (mu) is already known,
it can be passed to avoid calculating it again via weighted.mean().
Per default, undefined values and weights are silently removed.
}

