% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weighted_t_test}
\alias{weighted_t_test}
\title{weighted_t_test}
\usage{
weighted_t_test(x, w, mu, conf.level = 0.95, alternative = "two.sided",
  na.rm = TRUE, effective = TRUE)
}
\arguments{
\item{x}{array of input values or measurements}

\item{w}{array of weights}

\item{mu}{mean to test the distribution of x against (0)}

\item{conf.level}{the confidence-level (0.95)}

\item{alternative}{alternative to use for testing (two.sided)}

\item{na.rm}{remove undefined values and weights (TRUE)}

\item{effective}{use the effective sample size (TRUE)}
}
\description{
This function performes a weighted t-test of provided values (x)
and weights (w) against the reference (mu). Per default, all
undefined values in x are removed and those with undefined weight are
also ignored. The effective sample size is used instead of the length
of the vector (x) to compensate for largely different weights.
}

